reorg selfie products
