+++ android/src/main/kotlin/com/smileidentity/flutter/products/selfie/SmileIDSmartSelfieEnrollment.kt
@@ -1,9 +1,9 @@
-package com.smileidentity.flutter
+package com.smileidentity.flutter.products.selfie
 
 import android.content.Context
 import androidx.compose.runtime.Composable
 import com.smileidentity.SmileID
-import com.smileidentity.compose.SmartSelfieAuthentication
+import com.smileidentity.compose.SmartSelfieEnrollment
 import com.smileidentity.flutter.views.SmileSelfieComposablePlatformView
 import com.smileidentity.util.randomUserId
 import io.flutter.plugin.common.BinaryMessenger
@@ -12,20 +12,20 @@ import io.flutter.plugin.platform.PlatformView
 import io.flutter.plugin.platform.PlatformViewFactory
 import kotlinx.collections.immutable.toImmutableMap
 
-internal class SmileIDSmartSelfieAuthentication private constructor(
+internal class SmileIDSmartSelfieEnrollment private constructor(
     context: Context,
     viewId: Int,
     messenger: BinaryMessenger,
     args: Map<String, Any?>,
 ) : SmileSelfieComposablePlatformView(context, VIEW_TYPE_ID, viewId, messenger, args) {
     companion object {
-        const val VIEW_TYPE_ID = "SmileIDSmartSelfieAuthentication"
+        const val VIEW_TYPE_ID = "SmileIDSmartSelfieEnrollment"
     }
 
     @Composable
     override fun Content(args: Map<String, Any?>) {
         val extraPartnerParams = args["extraPartnerParams"] as? Map<String, String> ?: emptyMap()
-        SmileID.SmartSelfieAuthentication(
+        SmileID.SmartSelfieEnrollment(
             userId = args["userId"] as? String ?: randomUserId(),
             allowNewEnroll = args["allowNewEnroll"] as? Boolean ?: false,
             allowAgentMode = args["allowAgentMode"] as? Boolean ?: false,
@@ -41,7 +41,7 @@ internal class SmileIDSmartSelfieAuthentication private constructor(
         PlatformViewFactory(StandardMessageCodec.INSTANCE) {
         override fun create(context: Context, viewId: Int, args: Any?): PlatformView {
             @Suppress("UNCHECKED_CAST")
-            return SmileIDSmartSelfieAuthentication(
+            return SmileIDSmartSelfieEnrollment(
                 context,
                 viewId,
                 messenger,
